requirements.txt
anki~=2.1.34
requests
colorama==0.3.7
ConfigArgParse==1.2.3
demjson==2.2.4
pyperclip==1.5.27
retrying==1.3.3
six==1.11.0
tqdm==4.19.5
PyExecJS
